+++ experimental/packages/opentelemetry-exporter-metrics-otlp-http/test/metricsHelper.ts
@@ -19,6 +19,9 @@ import {
   ObservableResult,
   Histogram,
   ValueType,
+  ObservableCounter,
+  ObservableGauge,
+  ObservableUpDownCounter,
 } from '@opentelemetry/api-metrics';
 import { Resource } from '@opentelemetry/resources';
 import * as assert from 'assert';
@@ -99,15 +102,17 @@ export function mockCounter(): Counter {
 export function mockObservableGauge(
   callback: (observableResult: ObservableResult) => void,
   name = 'double-observable-gauge'
-): void {
-  return meter.createObservableGauge(
+): ObservableGauge {
+  const observableGauge = meter.createObservableGauge(
     name,
-    callback,
     {
       description: 'sample observable gauge description',
       valueType: ValueType.DOUBLE,
     }
   );
+  observableGauge.addCallback(callback);
+
+  return observableGauge;
 }
 
 export function mockDoubleCounter(): Counter {
@@ -123,29 +128,33 @@ export function mockDoubleCounter(): Counter {
 export function mockObservableCounter(
   callback: (observableResult: ObservableResult) => void,
   name = 'double-observable-counter'
-): void {
-  meter.createObservableCounter(
+): ObservableCounter {
+  const observableCounter = meter.createObservableCounter(
     name,
-    callback,
     {
       description: 'sample observable counter description',
       valueType: ValueType.DOUBLE,
     }
   );
+  observableCounter.addCallback(callback);
+
+  return observableCounter;
 }
 
 export function mockObservableUpDownCounter(
   callback: (observableResult: ObservableResult) => void,
   name = 'double-up-down-observable-counter'
-): void {
-  meter.createObservableUpDownCounter(
+): ObservableUpDownCounter {
+  const observableUpDownCounter = meter.createObservableUpDownCounter(
     name,
-    callback,
     {
       description: 'sample observable up down counter description',
       valueType: ValueType.DOUBLE,
     },
   );
+  observableUpDownCounter.addCallback(callback);
+
+  return observableUpDownCounter;
 }
 
 export function mockHistogram(): Histogram {